feature: build custom role

+++ controllers/controllerFactory.js
@@ -3,28 +3,48 @@ import AppError from "../utils/appError.js";
 
 export const createOne = (Model, parentName) =>
   catchAsyncErrors(async (req, res, next) => {
-    const { _id: id } = req.currentUser;
+    if (parentName === "user") {
+      const { _id: id } = req.currentUser;
 
-    if (!id) {
-      return next(
-        new AppError(
-          "You cannot create new documents when logged out. Try logging in first.",
-        ),
-      );
+      if (!id) {
+        return next(
+          new AppError(
+            "You cannot create new documents when logged out. Try logging in first.",
+          ),
+        );
+      }
+
+      const parent = { user: id };
+
+      const newDoc = await Model.create({ ...req.body, ...parent });
+
+      return res.status(201).json({
+        status: "success",
+        data: {
+          newDoc,
+        },
+      });
     }
 
-    let parent = { user: id };
-
     if (parentName) {
-      parent = {
+      const parent = {
         [convertCase(parentName, "camel")]:
           req.params[`${convertCase(parentName, "camel")}Id`],
       };
+
+      const newDoc = await Model.create({ ...req.body, ...parent });
+
+      return res.status(201).json({
+        status: "success",
+        data: {
+          newDoc,
+        },
+      });
     }
 
-    const newDoc = await Model.create({ ...req.body, ...parent });
+    const newDoc = await Model.create({ ...req.body });
 
-    res.status(201).json({
+    return res.status(201).json({
       status: "success",
       data: {
         newDoc,
@@ -34,15 +54,58 @@ export const createOne = (Model, parentName) =>
 
 export const getAll = (Model, parentName) =>
   catchAsyncErrors(async (req, res, next) => {
-    let query = {};
+    if (parentName === "user") {
+      const { _id: id } = req.currentUser;
+
+      if (!id) {
+        return next(
+          new AppError(
+            "You cannot create new documents when logged out. Try logging in first.",
+          ),
+        );
+      }
+
+      const query = { user: id };
+
+      const docs = await Model.find(query);
+
+      if (!docs) {
+        return next(
+          new AppError(`Sorry, couldn't find any docs with that query.`, 404),
+        );
+      }
+
+      return res.status(200).json({
+        status: "success",
+        data: {
+          docs,
+        },
+      });
+    }
+
     if (parentName) {
-      query = {
+      const query = {
         [convertCase(parentName, "camel")]:
           req.params[`${convertCase(parentName, "camel")}Id`],
       };
+
+      const docs = await Model.find(query);
+
+      if (!docs) {
+        return next(
+          new AppError(`Sorry, couldn't find any docs with that query.`, 404),
+        );
+      }
+
+      return res.status(200).json({
+        status: "success",
+        data: {
+          docs,
+        },
+      });
     }
 
-    const docs = await Model.find(query);
+    const docs = await Model.find();
 
     if (!docs) {
       return next(
@@ -50,7 +113,7 @@ export const getAll = (Model, parentName) =>
       );
     }
 
-    res.status(200).json({
+    return res.status(200).json({
       status: "success",
       data: {
         docs,
@@ -60,16 +123,70 @@ export const getAll = (Model, parentName) =>
 
 export const getOne = (Model, childName, parentName) =>
   catchAsyncErrors(async (req, res, next) => {
-    let query = { _id: req.params[`${convertCase(childName, "camel")}Id`] };
+    if (parentName === "user") {
+      const { _id: id } = req.currentUser;
+
+      if (!id) {
+        return next(
+          new AppError(
+            "You cannot create new documents when logged out. Try logging in first.",
+          ),
+        );
+      }
+
+      const query = {
+        user: id,
+        _id: req.params[`${convertCase(childName, "camel")}Id`],
+      };
+
+      const doc = await Model.findOne(query);
+
+      if (!doc) {
+        return next(
+          new AppError(
+            `Sorry, couldn't find any document with that ID: ${req.params[`${convertCase(childName, "camel")}Id`]}`,
+            404,
+          ),
+        );
+      }
+
+      return res.status(200).json({
+        status: "success",
+        data: {
+          doc,
+        },
+      });
+    }
+
     if (parentName) {
-      query = {
+      const query = {
         _id: req.params[`${convertCase(childName, "camel")}Id`],
         [convertCase(parentName, "camel")]:
           req.params[`${convertCase(parentName, "camel")}Id`],
       };
+
+      const doc = await Model.findOne(query);
+
+      if (!doc) {
+        return next(
+          new AppError(
+            `Sorry, couldn't find any document with that ID: ${req.params[`${convertCase(childName, "camel")}Id`]}`,
+            404,
+          ),
+        );
+      }
+
+      return res.status(200).json({
+        status: "success",
+        data: {
+          doc,
+        },
+      });
     }
 
-    const doc = await Model.findOne(query);
+    const doc = await Model.findOne({
+      _id: req.params[`${convertCase(childName, "camel")}Id`],
+    });
 
     if (!doc) {
       return next(
@@ -80,7 +197,7 @@ export const getOne = (Model, childName, parentName) =>
       );
     }
 
-    res.status(200).json({
+    return res.status(200).json({
       status: "success",
       data: {
         doc,
@@ -90,14 +207,76 @@ export const getOne = (Model, childName, parentName) =>
 
 export const updateOne = (Model, childName, parentName) =>
   catchAsyncErrors(async (req, res, next) => {
-    let query = { _id: req.params[`${convertCase(childName, "camel")}Id`] };
+    if (parentName === "user") {
+      const { _id: id } = req.currentUser;
+
+      if (!id) {
+        return next(
+          new AppError(
+            "You cannot create new documents when logged out. Try logging in first.",
+          ),
+        );
+      }
+
+      const query = {
+        user: id,
+        _id: req.params[`${convertCase(childName, "camel")}Id`],
+      };
+
+      const updatedDoc = await Model.findOneAndUpdate(query, req.body, {
+        new: true,
+        runValidators: true,
+      });
+
+      if (!updatedDoc) {
+        return next(
+          new AppError(
+            `Sorry, couldn't update because we find any document with that ID: ${req.params[`${convertCase(childName, "camel")}Id`]}`,
+            404,
+          ),
+        );
+      }
+
+      return res.status(200).json({
+        status: "success",
+        data: {
+          updatedDoc,
+        },
+      });
+    }
+
     if (parentName) {
-      query = {
+      const query = {
         _id: req.params[`${convertCase(childName, "camel")}Id`],
         [convertCase(parentName, "camel")]:
           req.params[`${convertCase(parentName, "camel")}Id`],
       };
+
+      const updatedDoc = await Model.findOneAndUpdate(query, req.body, {
+        new: true,
+        runValidators: true,
+      });
+
+      if (!updatedDoc) {
+        return next(
+          new AppError(
+            `Sorry, couldn't update because we find any document with that ID: ${req.params[`${convertCase(childName, "camel")}Id`]}`,
+            404,
+          ),
+        );
+      }
+
+      return res.status(200).json({
+        status: "success",
+        data: {
+          updatedDoc,
+        },
+      });
     }
+
+    const query = {
+      _id: req.params[`${convertCase(childName, "camel")}Id`],
+    };
 
     const updatedDoc = await Model.findOneAndUpdate(query, req.body, {
       new: true,
@@ -113,7 +292,7 @@ export const updateOne = (Model, childName, parentName) =>
       );
     }
 
-    res.status(200).json({
+    return res.status(200).json({
       status: "success",
       data: {
         updatedDoc,
@@ -123,14 +302,66 @@ export const updateOne = (Model, childName, parentName) =>
 
 export const deleteOne = (Model, childName, parentName) =>
   catchAsyncErrors(async (req, res, next) => {
-    let query = { _id: req.params[`${convertCase(childName, "camel")}Id`] };
+    if (parentName === "user") {
+      const { _id: id } = req.currentUser;
+
+      if (!id) {
+        return next(
+          new AppError(
+            "You cannot create new documents when logged out. Try logging in first.",
+          ),
+        );
+      }
+
+      const query = {
+        user: id,
+        _id: req.params[`${convertCase(childName, "camel")}Id`],
+      };
+
+      const deletedDoc = await Model.findOneAndDelete(query);
+
+      if (!deletedDoc) {
+        return next(
+          new AppError(
+            `Sorry, couldn't delete because we find any document with that ID: ${childName}`,
+            404,
+          ),
+        );
+      }
+
+      return res.status(204).json({
+        status: "successful",
+        data: null,
+      });
+    }
+
     if (parentName) {
-      query = {
+      const query = {
         _id: req.params[`${convertCase(childName, "camel")}Id`],
         [convertCase(parentName, "camel")]:
           req.params[`${convertCase(parentName, "camel")}Id`],
       };
+
+      const deletedDoc = await Model.findOneAndDelete(query);
+
+      if (!deletedDoc) {
+        return next(
+          new AppError(
+            `Sorry, couldn't delete because we find any document with that ID: ${childName}`,
+            404,
+          ),
+        );
+      }
+
+      return res.status(204).json({
+        status: "successful",
+        data: null,
+      });
     }
+
+    const query = {
+      _id: req.params[`${convertCase(childName, "camel")}Id`],
+    };
 
     const deletedDoc = await Model.findOneAndDelete(query);
 
@@ -143,7 +374,7 @@ export const deleteOne = (Model, childName, parentName) =>
       );
     }
 
-    res.status(204).json({
+    return res.status(204).json({
       status: "successful",
       data: null,
     });
@@ -151,18 +382,44 @@ export const deleteOne = (Model, childName, parentName) =>
 
 export const deleteAll = (Model, parentName) =>
   catchAsyncErrors(async (req, res, next) => {
-    const { _id: id } = req.currentUser;
-    let query = { user: id };
+    if (parentName === "user") {
+      const { _id: id } = req.currentUser;
+
+      if (!id) {
+        return next(
+          new AppError(
+            "You cannot create new documents when logged out. Try logging in first.",
+          ),
+        );
+      }
+
+      const query = { user: id };
+
+      const deleteCount = await Model.deleteMany(query);
+
+      return res.status(200).json({
+        status: "success",
+        deleteCount,
+      });
+    }
+
     if (parentName) {
-      query = {
+      const query = {
         [convertCase(parentName, "camel")]:
           req.params[`${convertCase(parentName, "camel")}Id`],
       };
+
+      const deleteCount = await Model.deleteMany(query);
+
+      return res.status(200).json({
+        status: "success",
+        deleteCount,
+      });
     }
 
-    const deleteCount = await Model.deleteMany(query);
+    const deleteCount = await Model.deleteMany();
 
-    res.status(200).json({
+    return res.status(200).json({
       status: "success",
       deleteCount,
     });
